backend/requirements.txt
# Core Django and Wagtail
Django>=5.2,<5.3
wagtail>=7.0,<7.1

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# API Framework
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
djangorestframework-simplejwt>=5.3.0

# Forum
django-machina>=1.3.0

# Cache
django-redis>=5.4.0
redis>=5.0.0

# Image Processing
Pillow>=10.3.0
django-imagekit>=5.0.0

# Environment
python-decouple>=3.8
python-dotenv>=1.0.0

# API Clients
requests>=2.32.0
httpx>=0.27.0

# Utilities
django-filter>=24.2
django-taggit>=5.0.0
django-mptt>=0.16.0
fuzzywuzzy>=0.18.0

# Development Tools
django-debug-toolbar>=4.4.0
django-extensions>=3.2.3
ipython>=8.24.0

# Testing
pytest>=8.2.0
pytest-django>=4.8.0
pytest-cov>=5.0.0
factory-boy>=3.3.0

# Task Queue (optional)
celery>=5.4.0

# Security
django-csp==3.8
django-ratelimit==4.1.0
python-magic==0.4.27

# OAuth Authentication
django-allauth>=0.58.2

# Production Server
gunicorn>=22.0.0
uvicorn[standard]>=0.30.0
django-celery-beat>=2.6.0

# WebSockets (Channels)
channels>=4.1.0
channels-redis>=4.2.0
daphne>=4.1.0

# AI Integration
wagtail-ai>=1.0.0

# Security Testing (Development)
bandit>=1.7.5
safety>=2.3.0

# Request tracing and logging
django-request-id>=1.0.0
python-json-logger>=2.0.7

# Error tracking
sentry-sdk[django,celery]>=2.0.0

# Static file optimization
whitenoise[brotli]>=6.6.0